requirements.txt
google.generativeai
pillow
streamlit
google.ai.generativelanguage